+++ fabric/src/main/java/net/modgarden/barricade/fabric/BarricadeFabric.java
@@ -59,7 +59,9 @@ public class BarricadeFabric implements ModInitializer {
 					BarricadeItems.MOB_BARRIER,
 					BarricadeItems.PASSIVE_BARRIER,
 					BarricadeItems.HOSTILE_BARRIER,
-					BarricadeItems.CREATIVE_ONLY_BARRIER);
+					BarricadeItems.CREATIVE_ONLY_BARRIER,
+					BarricadeItems.CREATIVE_ONLY_LEVER,
+					BarricadeItems.CREATIVE_ONLY_BUTTON);
 		});
 	}
 }