+++ CHANGELOG.md
@@ -1,2 +1,6 @@
 ## Bugfixes
-- Rectify buggy implementation of Advanced Barriers.
+- typo: `condition` field (previously `conditions`) in `PredicateBarrierBlock`
+- fix: Don't cache registry accesses
+
+## Changes
+- Operator items can now only be placed in creative mode